Refactor backtests to share filter logic
